ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.27.0
openai>=1.50.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.10.3,<=2.10.4
PyYAML==6.0.1 